% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{adding_peaks_to_df}
\alias{adding_peaks_to_df}
\title{Get Peak Flows}
\usage{
adding_peaks_to_df(data, parallel, wy_month, ...)
}
\arguments{
\item{data}{data.frame}

\item{parallel}{whether to use future_map or not}

\item{...}{other stuff to pass to future_map}
}
\value{
a df with peaks added to the water year
}
\description{
Get Peak Flows
}
