% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{clean_hourly_dv_report}
\alias{clean_hourly_dv_report}
\title{clean up hourly report}
\usage{
clean_hourly_dv_report(pp, data, days, ...)
}
\arguments{
\item{days}{how manys days back}

\item{...}{extra stuff to pass to ww_hourlyUSGS}

\item{usgs_stats_dv}{a data.frame}
}
\value{
a data.frame
}
\description{
clean up hourly report
}
