% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_SNOTEL.R
\name{get_daily_snotel_report}
\alias{get_daily_snotel_report}
\title{Get SNOTEL reports}
\usage{
get_daily_snotel_report(data, choice_days, type)
}
\arguments{
\item{data}{previously created api call}

\item{choice_days}{how much time to go back}

\item{type}{whether 'daily' or 'monthly'}
}
\value{
a data.frame with values from SNOTEL report generator.
}
\description{
Get SNOTEL reports
}
