% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_SNOTEL.R
\name{hourlySNOTEL}
\alias{hourlySNOTEL}
\title{Hourly SNOTEL}
\usage{
hourlySNOTEL(procDV, sites = NULL, days = 7, parallel = FALSE, ...)
}
\arguments{
\item{procDV}{A previously created \link[whitewater]{batch_SNOTELdv} object}

\item{sites}{A vector of SNOTEL site locations, e.g. \code{c("311", "500")}}

\item{days}{A \code{numeric} input of days, e.g. 1 = 24 hrs.}

\item{parallel}{\code{logical} indicating whether to use future_map().}

\item{...}{arguments to pass on to \link[furrr]{future_map}.}
}
\value{
A data.frame with hourly SNOTEL data
}
\description{
This function gets hourly SNOTEL data from \url{https://wcc.sc.egov.usda.gov/reportGenerator/} website.
}
