% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{hr_USGS}
\alias{hr_USGS}
\title{Prep USGS Hourly}
\usage{
hr_USGS(data, options, type)
}
\arguments{
\item{data}{the original data.frame}

\item{options}{a list of API arguments}

\item{type}{character of API call}
}
\description{
Prep USGS Hourly
}
