% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{iv_error_codes}
\alias{iv_error_codes}
\title{Get error codes}
\usage{
iv_error_codes(x)
}
\description{
Used for instantneous API call where sometimes an error occurrs. We just
want to capture the description.
}
