% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{month_to_doy}
\alias{month_to_doy}
\title{water year to months}
\usage{
month_to_doy(wy_month, leap = FALSE)
}
\arguments{
\item{wy_month}{A numeric}

\item{leap}{Logical}
}
\value{
A numeric value
}
\description{
Change wy_month to doy.
}
