% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{monthUSGS}
\alias{monthUSGS}
\title{Month-Only Stats (USGS)}
\usage{
monthUSGS(procDV, sites = NULL, parallel = FALSE, ...)
}
\arguments{
\item{procDV}{A previously created \link[whitewater]{batch_USGSdv} object.}

\item{sites}{A \code{character} vector with NWIS site numbers.}

\item{parallel}{\code{logical} indicating whether to use future_map().}

\item{...}{arguments to pass on to \link[furrr]{future_map} and \link[whitewater]{batch_USGSdv}.}
}
\value{
A \code{data.frame}
}
\description{
This function uses the results of the \link[whitewater]{batch_USGSdv} object to
generate mean, maximum, median and standard deviation for month-only.
}
