% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{pad_zero_for_logging}
\alias{pad_zero_for_logging}
\title{Prepping for loggin}
\usage{
pad_zero_for_logging(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
values are padded if zero for paramater of interest
}
\description{
Prepping for loggin
}
