% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{peaks_USGS}
\alias{peaks_USGS}
\title{Prepping Peaks}
\usage{
peaks_USGS(site_no, wy_month)
}
\arguments{
\item{site_no}{}
}
\value{
a data.frame with instantaneous peaks from USGS
}
\description{
Prepping Peaks
}
