% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{plot_baseflow}
\alias{plot_baseflow}
\title{Plot Water Year Base-flow and Total Flow}
\usage{
plot_baseflow(procDV, sites = NULL, wy, parallel = FALSE, ...)
}
\arguments{
\item{procDV}{A previously created \link[whitewater]{batch_USGSdv} object.}

\item{sites}{A \code{character} USGS NWIS site.}

\item{wy}{A \code{numeric} or \code{character} water year, e.g. 1990:2000, 1990, or c("1990", "2000").}

\item{parallel}{\code{logical} indicating whether to use future_map() if using \code{sites} argument.}

\item{...}{arguments to pass on to \link[furrr]{future_map}, \link[whitewater]{batch_USGSdv} and \link[lfstat]{baseflow}.}
}
\description{
Plot Water Year Base-flow and Total Flow
}
