% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{plot_USGSfdc}
\alias{plot_USGSfdc}
\title{USGS Flow Duration Curve}
\usage{
plot_USGSfdc(
  procDV,
  site = NULL,
  startDate = "",
  endDate = "",
  span_season = NULL
)
}
\arguments{
\item{procDV}{A previously created \link[whitewater]{batch_USGSdv} object.\code{optional}}

\item{site}{A USGS NWIS site. \code{optional}}

\item{startDate}{A \code{character} indicating the start date. minimum (default).}

\item{endDate}{A \code{character} indicating the end date. maximum (default)}

\item{span_season}{A \code{character} vector indicating a season, e.g. c('4-1', '6-30'). If NULL (default), yearly FDC will be produced.}
}
\value{
A ggplot. Works with plotly::ggplotly.
}
\description{
This function generates Flow Duration Curves.
}
