% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{prepping_reports}
\alias{prepping_reports}
\title{Prep Reports}
\usage{
prepping_reports(procDV, sites, parameterCd)
}
\arguments{
\item{parameterCd}{}
}
\value{
A prepped tibble meta data for api call
}
\description{
Prep Reports
}
