% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{prepping_USGSdv}
\alias{prepping_USGSdv}
\title{Prep USGS daily}
\usage{
prepping_USGSdv(site_no, parameterCd, start_date, end_date, statCd)
}
\arguments{
\item{statCd}{}
}
\value{
A tidied data frame with gage data dn info
}
\description{
Prep USGS daily
}
