% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_SNOTEL.R
\name{reportSNOTELmv}
\alias{reportSNOTELmv}
\title{Snotel Report Monthly}
\usage{
reportSNOTELmv(procDV, sites = NULL)
}
\arguments{
\item{procDV}{A previously created \link[whitewater]{batch_SNOTELdv} object}

\item{sites}{A vector of SNOTEL site locations, e.g. \code{c("311", "500")}}
}
\value{
A \code{data.frame}
}
\description{
This function gets monthly SNOTEL report data from \url{https://wcc.sc.egov.usda.gov/reportGenerator/} website. This includes
percentage of median from 1981-2010. Also included is current years data and the previous years data as well.
}
