% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{reportUSGSdv}
\alias{reportUSGSdv}
\title{USGS Report Daily}
\usage{
reportUSGSdv(procDV, sites = NULL, days = 10, parallel = FALSE, ...)
}
\arguments{
\item{procDV}{A previously created \link[whitewater]{batch_USGSdv} object.}

\item{sites}{A \code{character} USGS NWIS site.}

\item{days}{A \code{numeric} input of days.}

\item{parallel}{\code{logical} indicating whether to use future_map().}

\item{...}{arguments to pass on to \link[furrr]{future_map}.}
}
\description{
This function uses the \link[dataRetrieval]{readNWISstat} to gather daily
statistics like quantiles/percentiles. Be aware, the \code{current_daily_mean_flow} variable is calculated from the \link[whitewater]{hourlyUSGS}
by taking the daily mean of the hourly data. Thus, the \code{current_daily_mean_flow} will look different than the \link[whitewater]{hourlyUSGS} and
\href{https://waterdata.usgs.gov/nwis/rt}{USGS Current Water Data} site instantaneous values as it should.
}
