% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skewed}
\alias{skewed}
\title{Skew}
\usage{
skewed(x, na.rm = TRUE, type = 3)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{remove NA's}

\item{type}{numeric}
}
\value{
Skew from the 'psych' package
}
\description{
Skew
}
