% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_SNOTEL.R
\name{snotel_daily}
\alias{snotel_daily}
\title{Prep snotel daily}
\usage{
snotel_daily(md)
}
\arguments{
\item{md}{meta data for snotel sites}
}
\value{
a data frame with daily snotel results
}
\description{
Prep snotel daily
}
