% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_SNOTEL.R
\name{snotel_hourly}
\alias{snotel_hourly}
\title{Prepping SNOTEL hourly}
\usage{
snotel_hourly(md, choice_days)
}
\arguments{
\item{md}{data.frame}

\item{choice_days}{numeric}
}
\value{
a data.frame with site hourly values
}
\description{
Prepping SNOTEL hourly
}
