% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_SNOTEL.R
\name{snotel_report_custom}
\alias{snotel_report_custom}
\title{Change colnames snotel report}
\usage{
snotel_report_custom(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
a function for changing colnames for report
}
\description{
hijacked 'snotelr' package functions and manipulated
}
