% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_SNOTEL.R
\name{snotel_wild_custom}
\alias{snotel_wild_custom}
\title{Change colnames snotel}
\usage{
snotel_wild_custom(df)
}
\arguments{
\item{df}{A data.frame}
}
\value{
a function for changing colnames
}
\description{
hijacked 'snotelr' package functions and manipulated
}
