% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_SNOTEL.R
\name{snotel_wym}
\alias{snotel_wym}
\title{Prep SNOTEL wym}
\usage{
snotel_wym(md)
}
\arguments{
\item{md}{data.frame}
}
\value{
returns a data.frame with water year and month stats from report generator online
}
\description{
Prep SNOTEL wym
}
