% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_SNOTEL.R
\name{snotel_yearly}
\alias{snotel_yearly}
\title{Prep SNOTEL Yearly}
\usage{
snotel_yearly(md)
}
\arguments{
\item{md}{data.frame}
}
\value{
a data.frame with yearly stats
}
\description{
Prep SNOTEL Yearly
}
