% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{usgs_stats_fun}
\alias{usgs_stats_fun}
\title{get daily stats}
\usage{
usgs_stats_fun(data, type)
}
\arguments{
\item{data}{data.frame}
}
\value{
a data.frame with daily or monthly stats from readNWISstat
}
\description{
get daily stats
}
