% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{usgs_statsdv_fun}
\alias{usgs_statsdv_fun}
\title{get daily stats}
\usage{
usgs_statsdv_fun(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
a data.frame with daily stats from readNWISstat
}
\description{
get daily stats
}
