% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_SNOTEL.R
\name{ww_dvSNOTEL}
\alias{ww_dvSNOTEL}
\title{Process SNOTEL daily values}
\usage{
ww_dvSNOTEL(
  sites,
  parameterCd = "WTEQ",
  start_date = "",
  end_date = "",
  statCd = "MEAN",
  value_type = "value",
  depth = NULL,
  wy_month = 10,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{sites}{A vector of SNOTEL site locations, e.g. \code{c("311", "500")}}

\item{parallel}{\code{logical} indicating whether to use future_map().}

\item{...}{arguments to pass on to \link[furrr]{future_map}.}
}
\value{
A \code{tibble} with user defined SNOTEL metrics.
}
\description{
This function gets daily snotel data from \url{https://wcc.sc.egov.usda.gov/reportGenerator/} website.
}
