% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{ww_dvUSGS}
\alias{ww_dvUSGS}
\title{Process USGS daily values}
\usage{
ww_dvUSGS(
  sites,
  parameterCd = "00060",
  start_date = "",
  end_date = "",
  statCd = "00003",
  parallel = FALSE,
  wy_month = 10,
  ...
)
}
\arguments{
\item{sites}{A vector of USGS NWIS sites}

\item{parameterCd}{A USGS code for metric, default is "00060".}

\item{start_date}{A character of date format, e.g. \code{"1990-09-01"}}

\item{end_date}{A character of date format, e.g. \code{"1990-09-01"}}

\item{statCd}{character USGS statistic code. This is usually 5 digits. Daily mean (00003) is the default.}

\item{parallel}{\code{logical} indicating whether to use future_map().}

\item{wy_month}{\code{numeric} indicating the start month of the water year. e.g. 10 (default).}

\item{...}{arguments to pass on to \link[furrr]{future_map}.}
}
\value{
A \code{tibble} with daily mean flow and added meta-data.
}
\description{
This function is basically a wrapper around \link[dataRetrieval]{readNWISdv} but includes
added variables like water year, lat/lon, station name, altitude and tidied dates.
}
