% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{ww_floorIVUSGS}
\alias{ww_floorIVUSGS}
\title{Floor IV USGS}
\usage{
ww_floorIVUSGS(
  procDV,
  sites = NULL,
  parameterCd = NULL,
  options = wwOptions(),
  parallel = FALSE,
  ...
)
}
\arguments{
\item{procDV}{A previously created \link[whitewater]{ww_dvUSGS} object.}

\item{sites}{A \code{vector} of USGS NWIS sites (optional).}

\item{parameterCd}{A USGS code parameter code, only if using \code{sites} argument.}

\item{options}{A \link[whitewater]{wwOptions} call.}

\item{parallel}{\code{logical} indicating whether to use future_map().}

\item{...}{arguments to pass on to \link[furrr]{future_map}.}
}
\value{
A \code{tibble} with a user defined interval time step.
}
\description{
This function generates instantaneous NWIS data from \url{https://waterservices.usgs.gov/nwis/iv/}
and then floors to a user defined interval with \link[whitewater]{wwOptions}
('1 hour' is default) by taking the mean.
}
\note{
For performance reasons, with multi-site retrievals you may
retrieve data since October 1, 2007 only. If a previously created \link[whitewater]{ww_dvUSGS} object is not used then the user needs to
provide a \code{sites} vector. This will run \link[whitewater]{ww_dvUSGS} in the background.
}
