% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_SNOTEL.R
\name{ww_monthSNOTEL}
\alias{ww_monthSNOTEL}
\title{Month-Only Stats (SNOTEL)}
\usage{
ww_monthSNOTEL(procDV)
}
\arguments{
\item{procDV}{A previously created \link[whitewater]{batch_SNOTELdv} object}
}
\value{
A data.frame
}
\description{
This function uses a \link[whitewater]{batch_SNOTELdv} object to generate
month only statistics for snow water equivalent and snow depth.
}
