% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{ww_peakUSGS}
\alias{ww_peakUSGS}
\title{Get Peak Flows}
\usage{
ww_peakUSGS(sites, parallel, wy_month = 10, verbose = TRUE, ...)
}
\arguments{
\item{sites}{A vector of USGS NWIS sites}

\item{parallel}{\code{logical} indicating whether to use future_map().}

\item{wy_month}{\code{numeric} indicating the start month of the water year. e.g. 10 (default).}

\item{verbose}{\code{logical} for printing information. TRUE (default).}

\item{...}{arguments to pass on to \link[furrr]{future_map}.}
}
\value{
a \code{tibble} with peaks by water year
}
\description{
Get Peak Flows
}
