% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{ww_reportUSGSav}
\alias{ww_reportUSGSav}
\title{USGS Report Annually}
\usage{
ww_reportUSGSav(
  procDV,
  sites = NULL,
  parameterCd = NULL,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{procDV}{A previously created \link[whitewater]{ww_dvUSGS} object.}

\item{sites}{A \code{character} USGS NWIS site.}

\item{parameterCd}{A USGS code parameter code, only if using \code{sites} argument.}

\item{parallel}{\code{logical} indicating whether to use future_map().}

\item{...}{arguments to pass on to \link[furrr]{future_map}.}
}
\description{
Uses \link[dataRetrieval]{readNWISstat} to gather monthly mean values. Furthermore,
monthly quantiles are generated similar to \link[whitewater]{ww_reportUSGSdv}.
}
