% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{ww_reportUSGSdv}
\alias{ww_reportUSGSdv}
\title{USGS Report Daily}
\usage{
ww_reportUSGSdv(procDV, sites = NULL, days = 10, parallel = FALSE, ...)
}
\arguments{
\item{procDV}{A previously created \link[whitewater]{ww_dvUSGS} object.}

\item{sites}{A \code{character} USGS NWIS site.}

\item{days}{A \code{numeric} input of days.}

\item{parallel}{\code{logical} indicating whether to use future_map().}

\item{...}{arguments to pass on to \link[furrr]{future_map}.}
}
\description{
This function uses the \link[dataRetrieval]{readNWISstat} to gather daily
statistics like quantiles/percentiles. Be aware, the parameter values ('Flow', 'Wtemp', etc) are calculated from the \link[whitewater]{ww_hourlyUSGS}
function by taking the daily mean of the hourly data. Thus, the instantaneous values will look different than the daily mean values, as it should.
}
