% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{ww_reportUSGSmv}
\alias{ww_reportUSGSmv}
\title{USGS Report Monthly}
\usage{
ww_reportUSGSmv(procDV, sites = NULL)
}
\arguments{
\item{procDV}{A previously created \link[whitewater]{ww_dvUSGS} object.}

\item{sites}{A \code{character} USGS NWIS site.}
}
\description{
Uses \link[dataRetrieval]{readNWISstat} to gather monthly mean flows. Furthermore,
monthly quantiles are generated similar to \link[whitewater]{ww_reportUSGSdv}.
}
