% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{ww_statsNWIS}
\alias{ww_statsNWIS}
\title{NWIS stats}
\usage{
ww_statsNWIS(
  procDV,
  sites = NULL,
  temporalFilter = "daily",
  parameter_cd = NULL,
  days = 10,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{procDV}{A previously created \link[whitewater]{ww_dvUSGS} object.}

\item{sites}{A \code{character} USGS NWIS site.}

\item{temporalFilter}{A \code{character} for the stat summary window, e.g. 'daily' (default), 'monthly', 'yearly'.}

\item{parameter_cd}{A USGS code parameter code, only if using \code{sites} argument.}

\item{days}{A \code{numeric} input of days to go back from today (only needed if using .temporalFilter = 'daily').}

\item{parallel}{\code{logical} indicating whether to use future_map().}

\item{...}{arguments to pass on to \link[furrr]{future_map}.}
}
\description{
This function uses the \link[dataRetrieval]{readNWISstat} to gather daily
statistics like quantiles/percentiles.
}
\note{
Be aware, the parameter values ('Flow', 'Wtemp', etc) are calculated from the \link[whitewater]{ww_floorIVUSGS}
function by taking the daily mean of the hourly data. Thus, the instantaneous values will look different than the daily mean values, as it should.
The \code{.temporalFilter} argument is used to generate the window of percentiles.
}
