% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{ww_wyUSGS}
\alias{ww_wyUSGS}
\title{Water Year Stats (USGS)}
\usage{
ww_wyUSGS(procDV, sites = NULL, parallel = FALSE, ...)
}
\arguments{
\item{procDV}{A previously created \link[whitewater]{ww_dvUSGS} object.}

\item{sites}{A \code{character} vector with NWIS site numbers.}

\item{parallel}{\code{logical} indicating whether to use future_map().}

\item{...}{arguments to pass on to \link[furrr]{future_map} and/or \link[whitewater]{ww_dvUSGS}.}
}
\description{
This function uses the results of the \link[whitewater]{ww_dvUSGS} object to
generate mean, maximum, median and standard deviation per water year.
}
