% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{wwfilterNULL}
\alias{wwfilterNULL}
\title{Filter Null List}
\usage{
wwfilterNULL(x)
}
\arguments{
\item{x}{A list whose NULL elements will be filtered}
}
\description{
Taken from leaflet filterNULL function
remove NULL elements from a list
}
