% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{wwOptions}
\alias{wwOptions}
\title{Options}
\usage{
wwOptions(
  date_range = "pfn",
  period = 11,
  dates = NULL,
  site_status = "all",
  ...
)
}
\arguments{
\item{date_range}{A \code{character}. Indicating how to call the API. 'pfn' = Period from now,
'date_range' = a date range, "recent" = the most recent value.}

\item{period}{A \code{numeric}. Return all values from a period from now.}

\item{dates}{A \code{vector}. Return all values within an absolute date range (start and end dates)}

\item{site_status}{A \code{character} indicating site status. Example, 'all' = both active and inactive,
'active' = only active sites, 'inactive' = only inactive sites.}

\item{...}{other options used for options.}
}
\value{
A list with API options.
}
\description{
Options
}
\note{
A site is considered active if; it has collected time-series (automated) data within the last 183 days (6 months) or
it has collected discrete (manually collected) data within 397 days (13 months).
}
