% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_USGS.R
\name{ww_instantaneousUSGS}
\alias{ww_instantaneousUSGS}
\title{Instantaneous USGS}
\usage{
ww_instantaneousUSGS(
  procDV,
  sites = NULL,
  parameterCd = NULL,
  options = wwOptions(),
  parallel = FALSE,
  ...
)
}
\arguments{
\item{procDV}{A previously created \link[whitewater]{ww_dvUSGS} object.}

\item{sites}{A \code{vector} of USGS NWIS sites. \code{optional}}

\item{parameterCd}{A USGS code parameter code, only if using \code{sites} argument.}

\item{options}{A \link[whitewater]{wwOptions} call.}

\item{parallel}{\code{logical} indicating whether to use future_map().}

\item{...}{arguments to pass on to \link[furrr]{future_map}.}
}
\description{
This function generates Instantaneous NWIS data from
\url{https://waterservices.usgs.gov/nwis/iv/}..
}
\note{
For performance reasons, with multi-site retrievals you may retrieve data since October 1, 2007 only.
}
